auto crack - quick change of crack - autofind exebased on name and path to main dir

+++ Game.cs
@@ -13,8 +13,8 @@ namespace WpfApp3
         public string Path_Directory { get; set; }
         public string Type { get; set; }
         public string Background { get; set; }
-        public double Date_Added { get; set; }
-        public double Last_Played { get; set; }
+        public long Date_Added { get; set; }
+        public long Last_Played { get; set; }
         public int SteamAppid { get; set; }
         public float Playtime { get; set; }
         public string Logo { get; set; }
